Move Listing Handlers to own service
