+++ src/Core/Content/Product/SalesChannel/Listing/ProductListingFeaturesSubscriber.php
@@ -58,11 +58,6 @@ class ProductListingFeaturesSubscriber implements EventSubscriberInterface
      * @internal
      */
     public function __construct(
-        private readonly Connection $connection,
-        private readonly EntityRepository $optionRepository,
-        private readonly EntityRepository $sortingRepository,
-        private readonly SystemConfigService $systemConfigService,
-        private readonly EventDispatcherInterface $dispatcher,
         private readonly ListingFeatures $listingFeatures
     ) {
     }
@@ -111,17 +106,11 @@ class ProductListingFeaturesSubscriber implements EventSubscriberInterface
         $criteria = $event->getCriteria();
         $context = $event->getSalesChannelContext();
 
-        if (!$request->get('order')) {
-            $request->request->set('order', $this->getSystemDefaultSorting($context));
+        if($criteria->hasState(self::ALREADY_HANDLED)){
+            return;
         }
 
-        $criteria->addAssociation('options');
-
-        $this->handlePagination($request, $criteria, $event->getSalesChannelContext());
-
-        $this->handleFilters($request, $criteria, $context);
-
-        $this->handleSorting($request, $criteria, $context);
+        $this->listingFeatures->handleListingRequest($request, $criteria, $context);
     }
 
     public function handleSearchRequest(ProductSearchCriteriaEvent $event): void
@@ -137,6 +126,7 @@ class ProductListingFeaturesSubscriber implements EventSubscriberInterface
         $this->listingFeatures->handleSearchRequest($request, $criteria, $context);
     }
 
+    // Fixme: move to ListingFeatures.php
     public function handleResult(ProductListingResultEvent $event): void
     {
         Profiler::trace('product-listing::feature-subscriber', function () use ($event): void {
@@ -160,6 +150,7 @@ class ProductListingFeaturesSubscriber implements EventSubscriberInterface
         });
     }
 
+    // Fixme: Move to ListingFeatures.php
     public function removeScoreSorting(ProductListingResultEvent $event): void
     {
         $sortings = $event->getResult()->getAvailableSortings();
@@ -170,469 +161,5 @@ class ProductListingFeaturesSubscriber implements EventSubscriberInterface
         }
 
         $event->getResult()->setAvailableSortings($sortings);
-    }
-
-    private function handleFilters(Request $request, Criteria $criteria, SalesChannelContext $context): void
-    {
-        $criteria->addAssociation('manufacturer');
-
-        $filters = $this->getFilters($request, $context);
-
-        $aggregations = $this->getAggregations($request, $filters);
-
-        foreach ($aggregations as $aggregation) {
-            $criteria->addAggregation($aggregation);
-        }
-
-        foreach ($filters as $filter) {
-            if ($filter->isFiltered()) {
-                $criteria->addPostFilter($filter->getFilter());
-            }
-        }
-
-        $criteria->addExtension('filters', $filters);
-    }
-
-    /**
-     * @return array<Aggregation>
-     */
-    private function getAggregations(Request $request, FilterCollection $filters): array
-    {
-        $aggregations = [];
-
-        if ($request->get('reduce-aggregations') === null) {
-            foreach ($filters as $filter) {
-                $aggregations = array_merge($aggregations, $filter->getAggregations());
-            }
-
-            return $aggregations;
-        }
-
-        foreach ($filters as $filter) {
-            $excluded = $filters->filtered();
-
-            if ($filter->exclude()) {
-                $excluded = $excluded->blacklist($filter->getName());
-            }
-
-            foreach ($filter->getAggregations() as $aggregation) {
-                if ($aggregation instanceof FilterAggregation) {
-                    $aggregation->addFilters($excluded->getFilters());
-
-                    $aggregations[] = $aggregation;
-
-                    continue;
-                }
-
-                $aggregation = new FilterAggregation(
-                    $aggregation->getName(),
-                    $aggregation,
-                    $excluded->getFilters()
-                );
-
-                $aggregations[] = $aggregation;
-            }
-        }
-
-        return $aggregations;
-    }
-
-    private function handlePagination(Request $request, Criteria $criteria, SalesChannelContext $context): void
-    {
-        $limit = $this->getLimit($request, $context);
-
-        $page = $this->getPage($request);
-
-        $criteria->setOffset(($page - 1) * $limit);
-        $criteria->setLimit($limit);
-        $criteria->setTotalCountMode(Criteria::TOTAL_COUNT_MODE_EXACT);
-    }
-
-    private function handleSorting(Request $request, Criteria $criteria, SalesChannelContext $context): void
-    {
-        /** @var ProductSortingCollection $sortings */
-        $sortings = $criteria->getExtension('sortings') ?? new ProductSortingCollection();
-        $sortings->merge($this->getAvailableSortings($request, $context->getContext()));
-
-        $currentSorting = $this->getCurrentSorting($sortings, $request);
-
-        $criteria->addSorting(
-            ...$currentSorting->createDalSorting()
-        );
-
-        $criteria->addExtension('sortings', $sortings);
-    }
-
-    private function getCurrentSorting(ProductSortingCollection $sortings, Request $request): ProductSortingEntity
-    {
-        $key = $request->get('order');
-
-        $sorting = $sortings->getByKey($key);
-        if ($sorting !== null) {
-            return $sorting;
-        }
-
-        throw new ProductSortingNotFoundException($key);
-    }
-
-    private function getAvailableSortings(Request $request, Context $context): ProductSortingCollection
-    {
-        $criteria = new Criteria();
-        $criteria->setTitle('product-listing::load-sortings');
-        $availableSortings = $request->get('availableSortings');
-        $availableSortingsFilter = [];
-
-        if ($availableSortings) {
-            arsort($availableSortings, \SORT_DESC | \SORT_NUMERIC);
-            $availableSortingsFilter = array_keys($availableSortings);
-
-            $criteria->addFilter(new EqualsAnyFilter('key', $availableSortingsFilter));
-        }
-
-        $criteria
-            ->addFilter(new EqualsFilter('active', true))
-            ->addSorting(new FieldSorting('priority', 'DESC'));
-
-        /** @var ProductSortingCollection $sortings */
-        $sortings = $this->sortingRepository->search($criteria, $context)->getEntities();
-
-        if ($availableSortings) {
-            $sortings->sortByKeyArray($availableSortingsFilter);
-        }
-
-        return $sortings;
-    }
-
-    private function getSystemDefaultSorting(SalesChannelContext $context): string
-    {
-        return $this->systemConfigService->getString(
-            'core.listing.defaultSorting',
-            $context->getSalesChannel()->getId()
-        );
-    }
-
-    /**
-     * @return array<int, non-falsy-string>
-     */
-    private function collectOptionIds(ProductListingResultEvent $event): array
-    {
-        $aggregations = $event->getResult()->getAggregations();
-
-        /** @var TermsResult|null $properties */
-        $properties = $aggregations->get('properties');
-
-        /** @var TermsResult|null $options */
-        $options = $aggregations->get('options');
-
-        $options = $options ? $options->getKeys() : [];
-        $properties = $properties ? $properties->getKeys() : [];
-
-        return array_unique(array_filter([...$options, ...$properties]));
-    }
-
-    private function groupOptionAggregations(ProductListingResultEvent $event): void
-    {
-        $ids = $this->collectOptionIds($event);
-
-        if (empty($ids)) {
-            return;
-        }
-
-        $criteria = new Criteria($ids);
-        $criteria->setLimit(500);
-        $criteria->addAssociation('group');
-        $criteria->addAssociation('media');
-        $criteria->addFilter(new EqualsFilter('group.filterable', true));
-        $criteria->setTitle('product-listing::property-filter');
-        $criteria->addSorting(new FieldSorting('id', FieldSorting::ASCENDING));
-
-        $mergedOptions = new PropertyGroupOptionCollection();
-
-        $repositoryIterator = new RepositoryIterator($this->optionRepository, $event->getContext(), $criteria);
-        while (($result = $repositoryIterator->fetch()) !== null) {
-            /** @var PropertyGroupOptionCollection $entities */
-            $entities = $result->getEntities();
-
-            $mergedOptions->merge($entities);
-        }
-
-        // group options by their property-group
-        $grouped = $mergedOptions->groupByPropertyGroups();
-        $grouped->sortByPositions();
-        $grouped->sortByConfig();
-
-        $aggregations = $event->getResult()->getAggregations();
-
-        // remove id results to prevent wrong usages
-        $aggregations->remove('properties');
-        $aggregations->remove('configurators');
-        $aggregations->remove('options');
-        /** @var EntityCollection<Entity> $grouped */
-        $aggregations->add(new EntityResult('properties', $grouped));
-    }
-
-    private function addCurrentFilters(ProductListingResultEvent $event): void
-    {
-        $result = $event->getResult();
-
-        $filters = $result->getCriteria()->getExtension('filters');
-        if (!$filters instanceof FilterCollection) {
-            return;
-        }
-
-        foreach ($filters as $filter) {
-            $result->addCurrentFilter($filter->getName(), $filter->getValues());
-        }
-    }
-
-    /**
-     * @return list<string>
-     */
-    private function getManufacturerIds(Request $request): array
-    {
-        $ids = $request->query->get('manufacturer', '');
-        if ($request->isMethod(Request::METHOD_POST)) {
-            $ids = $request->request->get('manufacturer', '');
-        }
-
-        if (\is_string($ids)) {
-            $ids = explode('|', $ids);
-        }
-
-        /** @var list<string> $ids */
-        $ids = array_filter((array) $ids);
-
-        return $ids;
-    }
-
-    /**
-     * @return list<string>
-     */
-    private function getPropertyIds(Request $request): array
-    {
-        $ids = $request->query->get('properties', '');
-        if ($request->isMethod(Request::METHOD_POST)) {
-            $ids = $request->request->get('properties', '');
-        }
-
-        if (\is_string($ids)) {
-            $ids = explode('|', $ids);
-        }
-
-        /** @var list<string> $ids */
-        $ids = array_filter((array) $ids);
-
-        return $ids;
-    }
-
-    private function getLimit(Request $request, SalesChannelContext $context): int
-    {
-        $limit = $request->query->getInt('limit', 0);
-
-        if ($request->isMethod(Request::METHOD_POST)) {
-            $limit = $request->request->getInt('limit', $limit);
-        }
-
-        $limit = $limit > 0 ? $limit : $this->systemConfigService->getInt('core.listing.productsPerPage', $context->getSalesChannel()->getId());
-
-        return $limit <= 0 ? 24 : $limit;
-    }
-
-    private function getPage(Request $request): int
-    {
-        $page = $request->query->getInt('p', 1);
-
-        if ($request->isMethod(Request::METHOD_POST)) {
-            $page = $request->request->getInt('p', $page);
-        }
-
-        return $page <= 0 ? 1 : $page;
-    }
-
-    private function getFilters(Request $request, SalesChannelContext $context): FilterCollection
-    {
-        $filters = new FilterCollection();
-
-        $filters->add($this->getManufacturerFilter($request));
-        $filters->add($this->getPriceFilter($request));
-        $filters->add($this->getRatingFilter($request));
-        $filters->add($this->getShippingFreeFilter($request));
-        $filters->add($this->getPropertyFilter($request));
-
-        if (!$request->request->get('manufacturer-filter', true)) {
-            $filters->remove('manufacturer');
-        }
-
-        if (!$request->request->get('price-filter', true)) {
-            $filters->remove('price');
-        }
-
-        if (!$request->request->get('rating-filter', true)) {
-            $filters->remove('rating');
-        }
-
-        if (!$request->request->get('shipping-free-filter', true)) {
-            $filters->remove('shipping-free');
-        }
-
-        if (!$request->request->get('property-filter', true)) {
-            $filters->remove('properties');
-
-            if (\count($propertyWhitelist = $request->request->all(self::PROPERTY_GROUP_IDS_REQUEST_PARAM))) {
-                $filters->add($this->getPropertyFilter($request, $propertyWhitelist));
-            }
-        }
-
-        $event = new ProductListingCollectFilterEvent($request, $filters, $context);
-        $this->dispatcher->dispatch($event);
-
-        return $filters;
-    }
-
-    private function getManufacturerFilter(Request $request): Filter
-    {
-        $ids = $this->getManufacturerIds($request);
-
-        return new Filter(
-            'manufacturer',
-            !empty($ids),
-            [new EntityAggregation('manufacturer', 'product.manufacturerId', 'product_manufacturer')],
-            new EqualsAnyFilter('product.manufacturerId', $ids),
-            $ids
-        );
-    }
-
-    /**
-     * @param array<string>|null $groupIds
-     */
-    private function getPropertyFilter(Request $request, ?array $groupIds = null): Filter
-    {
-        $ids = $this->getPropertyIds($request);
-
-        $propertyAggregation = new TermsAggregation('properties', 'product.properties.id');
-
-        $optionAggregation = new TermsAggregation('options', 'product.options.id');
-
-        if ($groupIds) {
-            $propertyAggregation = new FilterAggregation(
-                'properties-filter',
-                $propertyAggregation,
-                [new EqualsAnyFilter('product.properties.groupId', $groupIds)]
-            );
-
-            $optionAggregation = new FilterAggregation(
-                'options-filter',
-                $optionAggregation,
-                [new EqualsAnyFilter('product.options.groupId', $groupIds)]
-            );
-        }
-
-        if (empty($ids)) {
-            return new Filter(
-                'properties',
-                false,
-                [$propertyAggregation, $optionAggregation],
-                new MultiFilter(MultiFilter::CONNECTION_OR, []),
-                [],
-                false
-            );
-        }
-
-        $grouped = $this->connection->fetchAllAssociative(
-            'SELECT LOWER(HEX(property_group_id)) as property_group_id, LOWER(HEX(id)) as id
-             FROM property_group_option
-             WHERE id IN (:ids)',
-            ['ids' => Uuid::fromHexToBytesList($ids)],
-            ['ids' => ArrayParameterType::STRING]
-        );
-
-        $grouped = FetchModeHelper::group($grouped);
-
-        $filters = [];
-        foreach ($grouped as $options) {
-            $options = array_column($options, 'id');
-
-            $filters[] = new MultiFilter(
-                MultiFilter::CONNECTION_OR,
-                [
-                    new EqualsAnyFilter('product.optionIds', $options),
-                    new EqualsAnyFilter('product.propertyIds', $options),
-                ]
-            );
-        }
-
-        return new Filter(
-            'properties',
-            true,
-            [$propertyAggregation, $optionAggregation],
-            new MultiFilter(MultiFilter::CONNECTION_AND, $filters),
-            $ids,
-            false
-        );
-    }
-
-    private function getPriceFilter(Request $request): Filter
-    {
-        $min = $request->get('min-price');
-        $max = $request->get('max-price');
-
-        $range = [];
-        if ($min !== null && $min >= 0) {
-            $range[RangeFilter::GTE] = $min;
-        }
-        if ($max !== null && $max >= 0) {
-            $range[RangeFilter::LTE] = $max;
-        }
-
-        return new Filter(
-            'price',
-            !empty($range),
-            [new StatsAggregation('price', 'product.cheapestPrice', true, true, false, false)],
-            new RangeFilter('product.cheapestPrice', $range),
-            [
-                'min' => (float) $request->get('min-price'),
-                'max' => (float) $request->get('max-price'),
-            ]
-        );
-    }
-
-    private function getRatingFilter(Request $request): Filter
-    {
-        $filtered = $request->get('rating');
-
-        return new Filter(
-            'rating',
-            $filtered !== null,
-            [
-                new FilterAggregation(
-                    'rating-exists',
-                    new MaxAggregation('rating', 'product.ratingAverage'),
-                    [new RangeFilter('product.ratingAverage', [RangeFilter::GTE => 0])]
-                ),
-            ],
-            new RangeFilter('product.ratingAverage', [
-                RangeFilter::GTE => (int) $filtered,
-            ]),
-            $filtered
-        );
-    }
-
-    private function getShippingFreeFilter(Request $request): Filter
-    {
-        $filtered = (bool) $request->get('shipping-free', false);
-
-        return new Filter(
-            'shipping-free',
-            $filtered === true,
-            [
-                new FilterAggregation(
-                    'shipping-free-filter',
-                    new MaxAggregation('shipping-free', 'product.shippingFree'),
-                    [new EqualsFilter('product.shippingFree', true)]
-                ),
-            ],
-            new EqualsFilter('product.shippingFree', true),
-            $filtered
-        );
     }
 }
